server/requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
argument==1.4.0
billiard==4.1.0
black==23.7.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.103.0
flower==2.0.1
h11==0.14.0
httptools==0.6.0
humanize==4.8.0
idna==3.4
kombu==5.3.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
ruff==0.0.286
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
